feat(streams): change API to better support use of streams
BREAKING CHANGE:
Main constructor function is now a named export, and has a different signature.
